requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
appnope==0.1.4
APScheduler==3.11.0
asttokens==3.0.0
attrs==25.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==44.0.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.6
fastjsonschema==2.21.1
filelock==3.16.1
Flask==2.3.2
Flask-Cors==4.0.1
Flask-Login==0.6.2
Flask-Mail==0.10.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
frozendict==2.4.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lxml==5.2.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.1
multitasking==0.0.11
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
peewee==3.17.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
platformdirs==4.3.6
plotly==5.20.0
pluggy==1.5.0
prompt_toolkit==3.0.50
proto-plus==1.25.0
protobuf==5.29.3
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
PyMySQL==1.0.2
pyparsing==3.2.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.1
qrcode==3.0
referencing==0.36.2
requests==2.31.0
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.22.3
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
sqlmodel==0.0.22
stack-data==0.6.3
starlette==0.41.3
stripe==11.1.0
tenacity==8.2.3
threadpoolctl==3.4.0
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.32.1
uvloop==0.21.0
virtualenv==20.26.6
watchfiles==1.0.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.1
Werkzeug==2.3.4
WTForms==3.1.2
yarg==0.1.9
yfinance==0.2.37
